Refactor processor decorator base class to be more generic
There really was very little commonality between single origin
and multiple origin processors, so instead the base class is now
a simple delegating implementation, where each method can be
overridden as needed.
