+++ subprojects/language-java/src/main/java/org/gradle/api/internal/tasks/compile/processing/DelegatingProcessor.java
@@ -1,0 +1,69 @@
+/*
+ * Copyright 2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.api.internal.tasks.compile.processing;
+
+import javax.annotation.processing.Completion;
+import javax.annotation.processing.ProcessingEnvironment;
+import javax.annotation.processing.Processor;
+import javax.annotation.processing.RoundEnvironment;
+import javax.lang.model.SourceVersion;
+import javax.lang.model.element.AnnotationMirror;
+import javax.lang.model.element.Element;
+import javax.lang.model.element.ExecutableElement;
+import javax.lang.model.element.TypeElement;
+import java.util.Set;
+
+/**
+ * A simple base class for annotation processors that decorate another one.
+ */
+abstract class DelegatingProcessor implements Processor {
+    private final Processor delegate;
+
+    DelegatingProcessor(Processor delegate) {
+        this.delegate = delegate;
+    }
+
+    @Override
+    public Set<String> getSupportedOptions() {
+        return delegate.getSupportedOptions();
+    }
+
+    @Override
+    public Set<String> getSupportedAnnotationTypes() {
+        return delegate.getSupportedAnnotationTypes();
+    }
+
+    @Override
+    public SourceVersion getSupportedSourceVersion() {
+        return delegate.getSupportedSourceVersion();
+    }
+
+    @Override
+    public void init(ProcessingEnvironment processingEnv) {
+        delegate.init(processingEnv);
+    }
+
+    @Override
+    public boolean process(Set<? extends TypeElement> annotations, RoundEnvironment roundEnv) {
+        return delegate.process(annotations, roundEnv);
+    }
+
+    @Override
+    public Iterable<? extends Completion> getCompletions(Element element, AnnotationMirror annotation, ExecutableElement member, String userText) {
+        return delegate.getCompletions(element, annotation, member, userText);
+    }
+}
